+++ lib/dynamic-height-gridview/dynamic_height_values.dart
@@ -2,9 +2,10 @@ import 'package:flutter/widgets.dart';
 
 List<double> getDynamicHeights(BuildContext context, int length) {
   List<double> dynamicHeight = [
+    MediaQuery.of(context).size.height * 0.7,
+    MediaQuery.of(context).size.height * 0.6,
     MediaQuery.of(context).size.height * 0.5,
-    MediaQuery.of(context).size.height * 0.45,
-    MediaQuery.of(context).size.height * 0.35,
+    MediaQuery.of(context).size.height * 0.4,
   ];
 
   List<double> result = [];