+++ CHANGELOG.md
@@ -1,3 +1,7 @@
-## 0.0.1
+# Changelog
 
-* TODO: Describe initial release.
+## [1.0.0] - 2024-12-10
+### Added
+- Initial release of Gridify.
+- Support for dynamic staggered grid layout.
+- Padding, scaling, color, and border-radius customizations.